dvc-service/requirements.txt
aiohttp>=3.11.13
fastapi>=0.115.11
python-dotenv>=1.0.1
python-multipart>=0.0.20
uvicorn>=0.34.0
